# api/main.go
package main

import (
	"log"
	"net/http"
	"strings"

	"github.com/aidashboard/api/config"
	"github.com/aidashboard/api/db"
	"github.com/aidashboard/api/handler"
	"github.com/aidashboard/api/service"
	"github.com/aidashboard/api/storage"
	"github.com/go-chi/chi/v5"
	chiMiddleware "github.com/go-chi/chi/v5/middleware"
)

func main() {
	cfg := config.Load()

	database, err := db.Connect(cfg.DatabaseURL)
	if err != nil {
		log.Fatalf("Failed to connect to database: %v", err)
	}
	defer database.Close()

	log.Println("Running migrations...")
	if err := db.RunMigrations(database); err != nil {
		log.Fatalf("Failed to run migrations: %v", err)
	}
	log.Println("Migrations complete")

	var minioStore *storage.MinioStorage
	if cfg.MinioConfigured() {
		minioStore, err = storage.NewMinioStorage(cfg)
		if err != nil {
			log.Fatalf("Failed to init MinIO storage: %v", err)
		}
		log.Println("MinIO storage ready")
	} else {
		log.Println("MinIO not configured, raw log upload disabled")
	}

	authH := handler.NewAuthHandler(database, cfg.JWTSecret)
	aiClient := service.NewAIClient()
	managedAgentClient := service.NewManagedAgentClient(cfg.ManagedAgentURL, cfg.ManagedAgentToken)
	reqH := handler.NewRequirementHandler(database, aiClient)
	taskH := handler.NewTaskHandler(database)
	sessionH := handler.NewSessionHandler(database, minioStore, aiClient)
	reportH := handler.NewReportHandler(database, cfg.ReportGeneratorURL)
	managedAgentH := handler.NewManagedAgentHandler(database, managedAgentClient)
	docH := handler.NewDocumentHandler(database)
	tokenH := handler.NewTokenHandler(database)
	teamH := handler.NewTeamHandler(database)
	followH := handler.NewFollowHandler(database)
	dashboardH := handler.NewDashboardHandler(database)

	r := chi.NewRouter()
	r.Use(chiMiddleware.Logger)
	r.Use(chiMiddleware.Recoverer)
	r.Use(corsMiddleware(cfg.CORSOrigin))

	r.Get("/health", func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(http.StatusOK)
		w.Write([]byte(`{"status":"ok"}`))
	})

	r.Post("/api/v1/auth/login", authH.Login)
	r.Post("/api/v1/auth/register", authH.Register)
	r.Get("/api/v1/users", authH.ListUsers)

	r.Route("/api/v1", func(r chi.Router) {
		r.Use(handler.AuthMiddleware(cfg.JWTSecret))

		r.Get("/auth/me", authH.Me)
		r.Get("/teams", authH.ListTeams)

		r.Route("/admin", func(r chi.Router) {
			r.Use(handler.AdminOnly)
			r.Put("/users/{id}", authH.AdminUpdateUser)
			r.Post("/users/{id}/reset-password", authH.AdminResetPassword)
		})

		r.Get("/requirements", reqH.List)
		r.Post("/requirements", reqH.Create)
		r.Get("/requirements/{id}", reqH.Get)
		r.Put("/requirements/{id}", reqH.Update)
		r.Delete("/requirements/{id}", reqH.Delete)
		r.Put("/requirements/{id}/restore", reqH.Restore)
		r.Get("/requirements/{id}/ac", reqH.GetAC)
		r.Post("/requirements/{id}/regenerate-ac", reqH.RegenerateAC)

		r.Get("/tasks", taskH.List)
		r.Post("/tasks", taskH.Create)
		r.Get("/tasks/{id}", taskH.Get)
		r.Put("/tasks/{id}", taskH.Update)
		r.Delete("/tasks/{id}", taskH.Delete)
		r.Put("/tasks/{id}/status", taskH.UpdateStatus)
		r.Put("/tasks/{id}/progress", taskH.UpdateProgress)
		r.Post("/tasks/{id}/dependencies", taskH.AddDependency)
		r.Delete("/tasks/{id}/dependencies/{dep_id}", taskH.RemoveDependency)

		r.Get("/follows", followH.List)
		r.Post("/follows", followH.Follow)
		r.Delete("/follows/{target_type}/{target_id}", followH.Unfollow)
		r.Get("/dashboard/follows", dashboardH.Follows)
		r.Get("/dashboard/risks", dashboardH.Risks)

		r.Post("/sessions/batch", sessionH.BatchUpload)
		r.Get("/sessions", sessionH.List)
		r.Get("/sessions/{id}", sessionH.Get)
		r.Get("/sessions/{id}/log", sessionH.DownloadLog)
		r.Put("/sessions/{id}/task", sessionH.UpdateTask)
		r.Put("/sessions/{id}/requirement", sessionH.UpdateRequirement)
		r.Delete("/sessions/{id}", sessionH.Withdraw)

		r.Get("/documents", docH.List)
		r.Post("/documents", docH.Create)
		r.Put("/documents/{id}", docH.Update)
		r.Delete("/documents/{id}", docH.Delete)

		r.Get("/reports", reportH.List)
		r.Get("/reports/today", reportH.GetOrCreateToday)
		r.Post("/reports/today/draft", reportH.GenerateTodayDraft)
		r.Post("/reports/today/managed-agent-runs", managedAgentH.StartDailyReportRun)
		r.Get("/reports/managed-agent-runs/{runId}", managedAgentH.GetDailyReportRun)
		r.Post("/reports/today/generate", reportH.GenerateToday)
		r.Get("/reports/{id}", reportH.Get)
		r.Put("/reports/{id}", reportH.Update)

		r.Get("/reports/team/members", reportH.ListTeamMemberReports)
		r.Get("/reports/team/today", reportH.GetTeamReportToday)
		r.Post("/reports/team/today/generate", reportH.GenerateTeamReport)
		r.Get("/reports/team", reportH.ListTeamReports)
		r.Put("/reports/team/{id}", reportH.UpdateTeamReport)

		r.Get("/tokens", tokenH.Aggregate)
		r.Get("/tokens/sessions", tokenH.ListSessionTokens)
		r.Get("/teams/activity", teamH.Activity)

		r.Get("/ai-assets/skills", managedAgentH.ListSkills)
		r.Get("/ai-assets/mcp", managedAgentH.ListMCPEntries)
		r.Post("/ai-assets/mcp", managedAgentH.CreateMCPEntry)
		r.Get("/ai-assets/agents", managedAgentH.ListMyAgents)
		r.Post("/ai-assets/agents", managedAgentH.CreateMyAgent)
		r.Put("/ai-assets/agents/{agentId}", managedAgentH.UpdateMyAgent)
		r.Post("/ai-assets/agents/{agentId}/runs", managedAgentH.StartAgentRun)
		r.Get("/ai-assets/agent-runs", managedAgentH.ListAgentRuns)
		r.Get("/ai-assets/agent-runs/{runId}", managedAgentH.GetAgentRun)
	})

	log.Printf("Starting API server on :%s", cfg.Port)
	if err := http.ListenAndServe(":"+cfg.Port, r); err != nil {
		log.Fatalf("Server failed: %v", err)
	}
}

func corsMiddleware(origin string) func(http.Handler) http.Handler {
	allowedOrigins := map[string]bool{}
	defaultOrigin := ""
	for _, item := range strings.Split(origin, ",") {
		item = strings.TrimSpace(item)
		if item != "" {
			if defaultOrigin == "" {
				defaultOrigin = item
			}
			allowedOrigins[item] = true
		}
	}

	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			requestOrigin := r.Header.Get("Origin")
			if allowedOrigins["*"] {
				w.Header().Set("Access-Control-Allow-Origin", "*")
			} else if allowedOrigins[requestOrigin] {
				w.Header().Set("Access-Control-Allow-Origin", requestOrigin)
				w.Header().Set("Vary", "Origin")
			} else if defaultOrigin != "" && requestOrigin == "" {
				w.Header().Set("Access-Control-Allow-Origin", defaultOrigin)
			}
			w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
			w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
			if r.Method == "OPTIONS" {
				w.WriteHeader(http.StatusOK)
				return
			}
			next.ServeHTTP(w, r)
		})
	}
}